Exclude ".." and "." from hasHiddenAncestorDirectory
