+++ src/test/resources/import-files/simple-realm/80_invalid_import_property.json
@@ -1,0 +1,3 @@
+{
+  "invalid": "value"
+}
