+++ src/test/resources/import-files/simple-realm/81_invalid_json.json
@@ -1,0 +1,2 @@
+{
+  "invalid": ""
